chore: internal refactor of loader to be structured like new mutation methods
