+++ packages/entity/src/EntityLoader.ts
@@ -1,14 +1,11 @@
 import AuthorizationResultBasedEntityLoader from './AuthorizationResultBasedEntityLoader';
 import EnforcingEntityLoader from './EnforcingEntityLoader';
 import { IEntityClass } from './Entity';
-import EntityConfiguration from './EntityConfiguration';
 import EntityLoaderUtils from './EntityLoaderUtils';
-import EntityPrivacyPolicy, { EntityPrivacyPolicyEvaluationContext } from './EntityPrivacyPolicy';
+import EntityPrivacyPolicy from './EntityPrivacyPolicy';
 import { EntityQueryContext } from './EntityQueryContext';
 import ReadonlyEntity from './ReadonlyEntity';
 import ViewerContext from './ViewerContext';
-import EntityDataManager from './internal/EntityDataManager';
-import IEntityMetricsAdapter from './metrics/IEntityMetricsAdapter';
 
 /**
  * The primary interface for loading entities. All normal loads are batched,
@@ -18,6 +15,7 @@ export default class EntityLoader<
   TFields extends object,
   TID extends NonNullable<TFields[TSelectedFields]>,
   TViewerContext extends ViewerContext,
+  TViewerContext2 extends TViewerContext,
   TEntity extends ReadonlyEntity<TFields, TID, TViewerContext, TSelectedFields>,
   TPrivacyPolicy extends EntityPrivacyPolicy<
     TFields,
@@ -28,26 +26,9 @@ export default class EntityLoader<
   >,
   TSelectedFields extends keyof TFields,
 > {
-  private readonly utilsPrivate: EntityLoaderUtils<
-    TFields,
-    TID,
-    TViewerContext,
-    TEntity,
-    TPrivacyPolicy,
-    TSelectedFields
-  >;
-
   constructor(
-    private readonly viewerContext: TViewerContext,
+    private readonly viewerContext: TViewerContext2,
     private readonly queryContext: EntityQueryContext,
-    private readonly privacyPolicyEvaluationContext: EntityPrivacyPolicyEvaluationContext<
-      TFields,
-      TID,
-      TViewerContext,
-      TEntity,
-      TSelectedFields
-    >,
-    private readonly entityConfiguration: EntityConfiguration<TFields>,
     private readonly entityClass: IEntityClass<
       TFields,
       TID,
@@ -56,23 +37,7 @@ export default class EntityLoader<
       TPrivacyPolicy,
       TSelectedFields
     >,
-    private readonly entitySelectedFields: TSelectedFields[] | undefined,
-    private readonly privacyPolicy: TPrivacyPolicy,
-    private readonly dataManager: EntityDataManager<TFields>,
-    protected readonly metricsAdapter: IEntityMetricsAdapter,
-  ) {
-    this.utilsPrivate = new EntityLoaderUtils(
-      this.viewerContext,
-      this.queryContext,
-      this.privacyPolicyEvaluationContext,
-      this.entityConfiguration,
-      this.entityClass,
-      this.entitySelectedFields,
-      this.privacyPolicy,
-      this.dataManager,
-      this.metricsAdapter,
-    );
-  }
+  ) {}
 
   /**
    * Enforcing entity loader. All loads through this loader are
@@ -103,14 +68,10 @@ export default class EntityLoader<
     TPrivacyPolicy,
     TSelectedFields
   > {
-    return new AuthorizationResultBasedEntityLoader(
-      this.queryContext,
-      this.entityConfiguration,
-      this.entityClass,
-      this.dataManager,
-      this.metricsAdapter,
-      this.utilsPrivate,
-    );
+    return this.viewerContext
+      .getViewerScopedEntityCompanionForClass(this.entityClass)
+      .getLoaderFactory()
+      .forLoad(this.queryContext, { previousValue: null, cascadingDeleteCause: null });
   }
 
   /**
@@ -125,6 +86,6 @@ export default class EntityLoader<
     TPrivacyPolicy,
     TSelectedFields
   > {
-    return this.utilsPrivate;
+    return this.withAuthorizationResults().utils;
   }
 }
