+++ packages/entity/src/__tests__/EntityMutator-test.ts
@@ -12,10 +12,10 @@ import {
 } from 'ts-mockito';
 import { v4 as uuidv4 } from 'uuid';
 
+import AuthorizationResultBasedEntityLoader from '../AuthorizationResultBasedEntityLoader';
 import EntityCompanionProvider from '../EntityCompanionProvider';
 import EntityConfiguration from '../EntityConfiguration';
 import EntityDatabaseAdapter from '../EntityDatabaseAdapter';
-import EntityLoader from '../EntityLoader';
 import EntityLoaderFactory from '../EntityLoaderFactory';
 import EntityLoaderUtils from '../EntityLoaderUtils';
 import {
@@ -577,7 +577,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id2),
       );
 
@@ -595,7 +594,6 @@ describe(EntityMutatorFactory, () => {
       const reloadedEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id2),
       );
       expect(reloadedEntity.getAllFields()).toMatchObject(updatedEntity.getAllFields());
@@ -632,7 +630,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, { previousValue: null, cascadingDeleteCause: null })
-          .withAuthorizationResults()
           .loadByIDAsync(id2),
       );
 
@@ -707,7 +704,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id2),
       );
 
@@ -780,7 +776,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id2),
       );
 
@@ -829,7 +824,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id1),
       );
 
@@ -845,7 +839,6 @@ describe(EntityMutatorFactory, () => {
       const reloadedEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id1),
       );
       expect(reloadedEntity.getAllFields()).toMatchObject(existingEntity.getAllFields());
@@ -884,7 +877,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id1),
       );
       expect(existingEntity).toBeTruthy();
@@ -897,7 +889,6 @@ describe(EntityMutatorFactory, () => {
         enforceAsyncResult(
           entityLoaderFactory
             .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-            .withAuthorizationResults()
             .loadByIDAsync(id1),
         ),
       ).rejects.toBeInstanceOf(Error);
@@ -937,7 +928,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id1),
       );
 
@@ -990,7 +980,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id1),
       );
 
@@ -1047,7 +1036,6 @@ describe(EntityMutatorFactory, () => {
       const existingEntity = await enforceAsyncResult(
         entityLoaderFactory
           .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-          .withAuthorizationResults()
           .loadByIDAsync(id1),
       );
 
@@ -1092,7 +1080,6 @@ describe(EntityMutatorFactory, () => {
     const entites1 = await enforceResultsAsync(
       entityLoaderFactory
         .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-        .withAuthorizationResults()
         .loadManyByFieldEqualingAsync('stringField', 'huh'),
     );
     expect(entites1).toHaveLength(1);
@@ -1107,7 +1094,6 @@ describe(EntityMutatorFactory, () => {
     const entities2 = await enforceResultsAsync(
       entityLoaderFactory
         .forLoad(viewerContext, queryContext, privacyPolicyEvaluationContext)
-        .withAuthorizationResults()
         .loadManyByFieldEqualingAsync('stringField', 'huh'),
     );
     expect(entities2).toHaveLength(2);
@@ -1170,17 +1156,16 @@ describe(EntityMutatorFactory, () => {
       },
     });
 
-    const entityLoaderMock =
-      mock<
-        EntityLoader<
-          SimpleTestFields,
-          string,
-          ViewerContext,
-          SimpleTestEntity,
-          SimpleTestEntityPrivacyPolicy,
-          keyof SimpleTestFields
-        >
-      >(EntityLoader);
+    const entityLoaderMock = mock<
+      AuthorizationResultBasedEntityLoader<
+        SimpleTestFields,
+        string,
+        ViewerContext,
+        SimpleTestEntity,
+        SimpleTestEntityPrivacyPolicy,
+        keyof SimpleTestFields
+      >
+    >(AuthorizationResultBasedEntityLoader);
     const entityLoaderUtilsMock =
       mock<
         EntityLoaderUtils<
@@ -1193,7 +1178,7 @@ describe(EntityMutatorFactory, () => {
         >
       >(EntityLoaderUtils);
     when(entityLoaderUtilsMock.constructEntity(anything())).thenReturn(fakeEntity);
-    when(entityLoaderMock.utils()).thenReturn(instance(entityLoaderUtilsMock));
+    when(entityLoaderMock.utils).thenReturn(instance(entityLoaderUtilsMock));
     const entityLoader = instance(entityLoaderMock);
 
     const entityLoaderFactoryMock =
@@ -1306,17 +1291,16 @@ describe(EntityMutatorFactory, () => {
       },
     });
 
-    const entityLoaderMock =
-      mock<
-        EntityLoader<
-          SimpleTestFields,
-          string,
-          ViewerContext,
-          SimpleTestEntity,
-          SimpleTestEntityPrivacyPolicy,
-          keyof SimpleTestFields
-        >
-      >(EntityLoader);
+    const entityLoaderMock = mock<
+      AuthorizationResultBasedEntityLoader<
+        SimpleTestFields,
+        string,
+        ViewerContext,
+        SimpleTestEntity,
+        SimpleTestEntityPrivacyPolicy,
+        keyof SimpleTestFields
+      >
+    >(AuthorizationResultBasedEntityLoader);
     const entityLoaderUtilsMock =
       mock<
         EntityLoaderUtils<
@@ -1329,7 +1313,7 @@ describe(EntityMutatorFactory, () => {
         >
       >(EntityLoaderUtils);
     when(entityLoaderUtilsMock.constructEntity(anything())).thenReturn(fakeEntity);
-    when(entityLoaderMock.utils()).thenReturn(instance(entityLoaderUtilsMock));
+    when(entityLoaderMock.utils).thenReturn(instance(entityLoaderUtilsMock));
     const entityLoader = instance(entityLoaderMock);
 
     const entityLoaderFactoryMock =
